Implemented Task 3.1

+++ minitorch/tensor_ops.py
@@ -273,8 +273,8 @@ def tensor_map(
         in_shape: Shape,
         in_strides: Strides,
     ) -> None:
-        out_index = np.zeros_like(out_shape)
-        in_index = np.zeros_like(in_shape)
+        out_index = np.zeros_like(out_shape, dtype=np.int32)
+        in_index = np.zeros_like(in_shape, dtype=np.int32)
 
         for i in range(len(out)):
             to_index(i, out_shape, out_index)
@@ -327,9 +327,9 @@ def tensor_zip(
         b_shape: Shape,
         b_strides: Strides,
     ) -> None:
-        out_index = np.zeros_like(out_shape)
-        a_index = np.zeros_like(a_shape)
-        b_index = np.zeros_like(b_shape)
+        out_index = np.zeros_like(out_shape, dtype=np.int32)
+        a_index = np.zeros_like(a_shape, dtype=np.int32)
+        b_index = np.zeros_like(b_shape, dtype=np.int32)
 
         for i in range(len(out)):
             to_index(i, out_shape, out_index)
@@ -370,7 +370,7 @@ def tensor_reduce(
         a_strides: Strides,
         reduce_dim: int,
     ) -> None:
-        out_index = np.zeros_like(out_shape)
+        out_index = np.zeros_like(out_shape, dtype=np.int32)
 
         for i in range(len(out)):
             to_index(i, out_shape, out_index)
